fix(r2-h): community_epoch u32→u64 to match DfrostLogEngine::current_epoch
Change Tier3PollMeta.community_epoch, set_tier3_poll_epoch, and
derive_beacon_seed signatures from u32 to u64, eliminating all
as-u64 casts at call sites. derive_beacon_seed now encodes the epoch
as 8 LE bytes instead of 4.

+++ src-tauri/src/community_voting_log.rs
@@ -602,7 +602,7 @@ impl VotingLog {
     /// request is issued after this call with the value that was read once.
     ///
     /// Returns `false` if the poll does not exist or is not Tier 3 (no-op).
-    pub fn set_tier3_poll_epoch(&mut self, poll_id: &PollId, epoch: u32) -> bool {
+    pub fn set_tier3_poll_epoch(&mut self, poll_id: &PollId, epoch: u64) -> bool {
         match self.polls.get_mut(poll_id) {
             Some(ps) => match ps.tier_state.as_tier3_mut() {
                 Some(t3) => {
